Move dummy customer to a separate configuration file
